backend/requirements.txt
# YOLOv8 차선별 차량 카운팅 시스템
# Python >=3.10,<3.12 권장 (검증 환경: miniconda py310_pt)
#
# GPU(CUDA) 사용 시 torch는 CUDA 빌드가 필요합니다.
# 예: pip install torch==2.4.1 --index-url https://download.pytorch.org/whl/cu121

# ===== 핵심 딥러닝 / 추적 =====
ultralytics>=8.4,<9.0
torch>=2.4,<3.0
torchvision>=0.19,<0.20

# ===== 컴퓨터 비전 =====
# opencv-python과 opencv-contrib-python을 동시에 설치하지 말 것 (심볼 충돌)
opencv-python>=4.10,<5.0

# ===== 수치 계산 / 시각화 =====
numpy>=1.24,<2.0
matplotlib>=3.7,<4.0
seaborn>=0.12,<1.0

# ===== 설정 파일 =====
PyYAML>=6.0,<7.0

# ===== API 서버 (FastAPI) =====
fastapi>=0.110,<1.0
uvicorn[standard]>=0.27,<1.0
python-multipart>=0.0.9        # 파일 업로드
aiofiles>=23.2,<25.0           # 비동기 파일 I/O
